+++ projects/volt/src/lib/components/toggle-group/toggle-group-item.ts
@@ -5,7 +5,11 @@ import {
   computed,
   input,
 } from '@angular/core';
-import { NgpToggleGroupItem, provideToggleGroupItemState } from 'ng-primitives/toggle-group';
+import {
+  injectToggleGroupState,
+  NgpToggleGroupItem,
+  provideToggleGroupItemState,
+} from 'ng-primitives/toggle-group';
 import { toggleGroupItemVariants, type ToggleGroupItemVariants } from './variants';
 
 @Component({
@@ -20,7 +24,7 @@ import { toggleGroupItemVariants, type ToggleGroupItemVariants } from './variant
     <button
       ngpToggleGroupItem
       [ngpToggleGroupItemValue]="value()"
-      [ngpToggleGroupItemDisabled]="disabled()"
+      [ngpToggleGroupItemDisabled]="isDisabled()"
       [class]="classes()"
       type="button"
     >
@@ -29,10 +33,15 @@ import { toggleGroupItemVariants, type ToggleGroupItemVariants } from './variant
   `,
 })
 export class VoltToggleGroupItem {
+  private readonly toggleGroupState = injectToggleGroupState();
+
   readonly value = input.required<string>();
   readonly disabled = input<boolean, unknown>(false, { transform: booleanAttribute });
   readonly variant = input<ToggleGroupItemVariants['variant']>('default');
   readonly size = input<ToggleGroupItemVariants['size']>('md');
+  protected readonly isDisabled = computed(
+    () => this.disabled() || this.toggleGroupState().disabled()
+  );
 
   protected readonly classes = computed(() =>
     toggleGroupItemVariants({ variant: this.variant(), size: this.size() })